Update executable to the lastest version.
Previous version needed to be run on top of Qwerty US International layout to make use of AltGr. However this version can (and must) be run on Qwerty US (non-International). This change was introduced with gaming in mind. When suspending the script to use Qwerty for games, AltGr is harder to remap compared to Left Alt.